requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
boto3==1.26.165
botocore==1.29.165
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.3.0
dj-database-url==2.0.0
Django==4.2.4
django-cors-headers==4.1.0
django-parler==2.3
django-rosetta==0.9.8
django-shortcuts==1.6
django-storages==1.13.2
exceptiongroup==1.1.3
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
jmespath==1.0.1
kombu==5.3.1
model-bakery==1.15.0
packaging==23.1
Pillow==10.0.0
pluggy==1.3.0
prompt-toolkit==3.0.39
psycopg2==2.9.7
psycopg2-binary==2.9.7
pytest==7.4.0
python-dateutil==2.8.2
python-decouple==3.8
requests==2.31.0
redis==4.6.0
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
stripe==5.5.0
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
weasyprint==56.1
whitenoise==6.5.0
